Moved JSONObject back to the root
