+++ BranchSDK-Samples/Windows/TestBed/MainPage.xaml.cpp
@@ -9,7 +9,7 @@
 #include <sstream>
 #include <BranchIO/AppInfo.h>
 #include <BranchIO/Branch.h>
-#include <BranchIO/Internal/JSONObject.h>
+#include <BranchIO/JSONObject.h>
 
 using namespace UWP101;
